# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody'
      meta: {}
      success:
        type: boolean
    type: object
  github_com_efangly_thanes-lims-backend_internal_adapters_http_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  internal_adapters_http_purchaseorder.MarkReceivedRequest:
    properties:
      expire_date:
        type: string
      lot_no:
        type: string
    required:
    - lot_no
    type: object
  internal_adapters_http_purchaseorder.POResponse:
    properties:
      id:
        type: string
      item_id:
        type: string
      order_date:
        type: string
      quantity:
        type: integer
      status:
        type: string
      vendor:
        type: string
    type: object
...
info:
  contact: {}
  description: REST API สำหรับ Thanes LIMS backend (Auth, Sample/Chain-of-Custody,
    Test Result, Equipment, Inventory/Purchase Order, Document, Environment, Notification,
    Audit)
  title: Thanes LIMS Backend API
  version: "1.0"
paths:
...
  /purchase-orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_adapters_http_purchaseorder.POResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: รายการใบสั่งซื้อทั้งหมด
      tags:
      - purchase-orders
  /purchase-orders/{id}:
    get:
      parameters:
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_purchaseorder.POResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: ดึงข้อมูลใบสั่งซื้อตาม ID
      tags:
      - purchase-orders
  /purchase-orders/{id}/approve:
    patch:
      parameters:
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_purchaseorder.POResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: อนุมัติใบสั่งซื้อ
      tags:
      - purchase-orders
  /purchase-orders/{id}/receive:
    patch:
      description: เพิ่มจำนวนคงคลังตามใบสั่งซื้อโดยอัตโนมัติ
      parameters:
      - description: Purchase Order ID
        in: path
        name: id
        required: true
        type: string
      - description: ล็อตที่รับเข้า
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_http_purchaseorder.MarkReceivedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_adapters_http_purchaseorder.POResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_efangly_thanes-lims-backend_internal_adapters_http_response.Envelope'
      security:
      - BearerAuth: []
      summary: บันทึกรับสินค้าตามใบสั่งซื้อ
      tags:
      - purchase-orders
...
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคและ JWT access token เช่น "Bearer eyJhbGciOi..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"